slightly more sensible test names
